feat: Add selectorName to ChecklistSection and ChecklistList
This commit introduces a `selectorName` field to both `ChecklistSection` and `ChecklistList` data models.

This new field allows for a different name to be displayed in the section and list selectors compared to the main title displayed for the section or list.

If `selectorName` is blank, the original `name` field will be used for display in the selectors.

The `checklist.json` has been updated to include `selectorName` where desired, primarily for shortening longer list and section names in the selection UI.