Make first task item active on checklist load
Previously, the first item in the checklist (regardless of type) was set as active when a checklist was loaded. This change modifies the behavior to set the first "TASK" type item as the active item.

If no "TASK" items are present, no item will be initially active (activeItemIndex will be -1).